commit -changes in DAO ArrayList to list,Change int to number in table

+++ WebContent/db/migration/airtickets.sql
@@ -1,4 +1,4 @@
-create table airlines(id int PRIMARY KEY,airlines_name varchar2(30) NOT NULL UNIQUE);
+create table airlines(id number(10) PRIMARY KEY,airlines_name varchar2(30) NOT NULL UNIQUE);
 
 ALTER TABLE airlines ADD (created_by int,created_time timestamp,updated_by int,updated_time timestamp);
 
@@ -6,7 +6,7 @@ create sequence seq_airlinesid
 start with 1
 increment by 1;
 
-create table passengers(id int CONSTRAINT pk_passangers_id primary key,name varchar2(20),email varchar2(30)not null,
+create table passengers(id number(10) CONSTRAINT pk_passangers_id primary key,name varchar2(20),email varchar2(30)not null,
 phonenumber long,username varchar2(20)not null,password varchar2(30) not null,
 created_date date);
 
@@ -18,7 +18,7 @@ create sequence seq_passengers_id
 start with 1
 increment by 1;
 
-create table airlines_flight(id int CONSTRAINT pk_airlines_id primary key,flight_name varchar2(30),flight_no varchar2(10),
+create table airlines_flight(id number(10) CONSTRAINT pk_airlines_id primary key,flight_name varchar2(30),flight_no varchar2(10),
 adult_seats int,child_seats int,adult_price int,child_price int,status char,flight_class varchar(30));
 
 alter table AIRLINES_FLIGHT ADD CONSTRAINT fk_flightname foreign key (flight_name) references airlines(id);
@@ -28,7 +28,7 @@ create sequence seq_airlines_id
 start with 1
 increment by 1;
 
-create table bookingairlines(id int primary key,airlines_id int CONSTRAINT fk_booking_airlines_id references airlines(id),
+create table bookingairlines(id number(10) primary key,airlines_id int CONSTRAINT fk_booking_airlines_id references airlines(id),
 adult_seats int,child_seats int,co_passangersname varchar2(40),
 price int,booking_date date,
 passengers_id int CONSTRAINT fk_booking_passengers_id references passengers(id));
